Move timeline into a view and break the components down
Also added the page level.
