+++ src/Components/Timeline/Container.component.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+
+import styled from "styled-components";
+
+const Container: React.FC = styled.div`
+    margin: 40px 0px;
+
+    display: flex;
+    flex-direction: column;
+    position: relative;
+`;
+
+export default Container;
