+++ src/Views/Timeline.view.tsx
@@ -1,0 +1,32 @@
+import React, { useContext } from "react";
+
+import { ThemeContext } from "styled-components";
+
+import Container from "@component/Timeline/Container.component";
+import TimelineLine from "@component/Timeline/TimelineLine.component";
+import VerticalTimelineElement from "@component/Timeline/VerticalElement.component";
+
+import { Work } from "@model/work";
+
+type Props = {
+    Worklist: Work[],
+};
+
+const Timeline: React.FC<Props> = (props: Props) => {
+    const theme = useContext(ThemeContext);
+
+    return (
+        <Container>
+            <TimelineLine BackgroundColor={theme.primaryHue} />
+            {props.Worklist.map((wl) => {
+                return (
+                    <VerticalTimelineElement
+                        key={wl.ID}
+                        Work={wl} />
+                );
+            })}
+        </Container>
+    );
+};
+
+export default Timeline;
