+++ tsconfig.json
@@ -17,10 +17,10 @@
         "baseUrl": "./",
         "paths": {
             "@api/*": ["./src/API/*"],
-            "@component/*": ["./src/Components/*"],
-            "@container/*": ["./src/Containers/*"],
             "@model/*": ["./src/Model/*"],
+            "@component/*": ["./src/Components/*"],
             "@view/*": ["./src/Views/*"],
+            "@page/*": ["./src/Pages/*"],
         },
         "types": [
             "chai",
